requirements.txt
numpy==2.2.6
scipy==1.16.3
astropy==7.2.0
skyfield==1.53
jplephem==2.23
pyswisseph==2.10.3.2
timezonefinder==6.0.1
pytz==2024.1
fastapi==0.122.0
uvicorn==0.38.0
better-profanity==0.7.0
nudenet==3.4.2
python-multipart==0.0.9
fasttext-wheel==0.9.2
Pillow==12.0.0
anyio==4.11.0
websockets==14.1
watchfiles==0.24.0
httptools==0.6.3
h11==0.16.0
